refactor: apply review suggestions to container and reporter imports
- PluginContainer implements Required<Plugin> and gains a name
- Collapse the checkMatch error handler into .catch
- Rename the filter parameter to checkPromise
- Remove the reporters barrel and the unused src/reporters.ts; import
  each reporter module directly and drop the extra imports-map entry
