+++ packages/todone/src/commands/run/impl.ts
@@ -1,6 +1,6 @@
 import { run } from "#/index";
 import { loadConfigFile } from "#/lib/config/load";
-import { cliReporter } from "#/lib/reporters";
+import { cliReporter } from "#/lib/reporters/cli";
 import { RunCommand } from "./index";
 
 export default async ({ onlyExpired, locale }: RunCommand) => {
