--- packages/todone/src/lib/reporters/index.ts
@@ -1,3 +1,0 @@
-export * from "./cli";
-export * from "./json";
-export * from "./silent";
