--- packages/todone/src/reporters.ts
@@ -1,1 +1,0 @@
-export * from "./lib/reporters";
